python_engine/requirements.txt
fastapi
uvicorn
pandas
ta
yfinance
urllib3<2.0
playwright
supabase
openai
python-dotenv
requests
